ЛР№3/requirements.txt
fastapi
uvicorn
httpx
pyjwt
bcrypt
python-multipart
